refactor: centralize exact limits
