+++ crates/loq_cli/src/lib.rs
@@ -10,6 +10,7 @@ mod baseline_shared;
 mod check;
 mod cli;
 mod config_edit;
+mod exact_limits;
 mod init;
 mod output;
 mod relax;
